Added an avatar component for iOS
